CEDS-647 Moved AuthAction to new file
